+++ bootique-jersey/src/test/java/io/bootique/jersey/ParamConvertersIT.java
@@ -32,9 +32,7 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
+import java.time.*;
 
 @BQTest
 public class ParamConvertersIT {
@@ -47,6 +45,18 @@ public class ParamConvertersIT {
             .module(b -> JerseyModule.extend(b).addResource(Resource.class))
             .module(jetty.moduleReplacingConnectors())
             .createRuntime();
+
+    @Test
+    public void testYearConverter() {
+        Response r = jetty.getTarget().path("year/2019").request().get();
+        JettyTester.assertOk(r).assertContent("[2019]");
+    }
+
+    @Test
+    public void testYearMonthConverter() {
+        Response r = jetty.getTarget().path("year-month/2019-05").request().get();
+        JettyTester.assertOk(r).assertContent("[2019-05]");
+    }
 
     @Test
     public void testDateConverter() {
@@ -68,6 +78,20 @@ public class ParamConvertersIT {
 
     @Path("/")
     public static class Resource {
+
+        @GET
+        @Produces(MediaType.TEXT_PLAIN)
+        @Path("year/{year}")
+        public String getY(@PathParam("year") Year year) {
+            return "[" + year + "]";
+        }
+
+        @GET
+        @Produces(MediaType.TEXT_PLAIN)
+        @Path("year-month/{year-month}")
+        public String getYM(@PathParam("year-month") YearMonth yearMonth) {
+            return "[" + yearMonth + "]";
+        }
 
         @GET
         @Produces(MediaType.TEXT_PLAIN)