+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #65 Upgrade to Jersey 2.38 / 3.0.9
 * #67 ParamConverter for LocalDate, LocalTime, LocalDateTime
+* #68 ParamConverter for Year, YearMonth
 
 ## 3.0.M1